feat: Add get_admin_name method to HelperFunctions
The code changes in `HelperFunctions.py` add a new method `get_admin_name` to the `HelperFunctions` class. This method retrieves the name of an admin from the `admins_collection` based on the admin's ID. If the admin is not found or an error occurs, the method returns "Unknown". This addition enhances the functionality of the `HelperFunctions` class by providing a way to get the name of an admin.