refactor: Sort final_data by createdDate in UserService
This commit refactors the UserService module to sort the final_data list by the "createdDate" field. By using the sorted() function with a custom key function, the data is sorted in descending order based on the "createdDate" value. This change ensures that the data is displayed in the correct order when retrieved from the database.