refactor: Add "source" field to meta_fields in EngagementService
This commit adds a new field, "source", to the `meta_fields` list in the `EngagementService` class. The "source" field is used to store the source of the engagement data. By including this field, we can track and analyze the origin of the data, providing valuable insights for reporting and analysis.

Closes #N/A