feat: Refactor WAService query to include "full" users
This commit refactors the query in the WAService.py file to include a check for "full" users. If the usersType is "full", the query will search for users with a completed profile. This change enhances the functionality of the WAService module by filtering out incomplete profiles from the list of users.